requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.0,<25.0
drf-spectacular>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<4.0
pytest>=7.0,<9.0
pytest-django>=4.5,<5.0
